delete old chat files
