+++ frontend/src/App.vue
@@ -15,7 +15,6 @@ import HelloWorld from "@/components/HelloWorld.vue";
       <nav class="bg-pan-left">
         <RouterLink to="/">Accueil</RouterLink>
         <RouterLink to="/chat">Chat</RouterLink>
-        <RouterLink to="/chat_new">Chat New</RouterLink>
         <RouterLink to="/friends">Amis</RouterLink>
         <RouterLink to="/game">Jeux</RouterLink>
         <RouterLink to="/account">Compte</RouterLink>
