+++ frontend/src/components/ChatNew/ChatAvailableRooms.vue
@@ -73,9 +73,8 @@ export default {
 };
 </script>
 <template>
-  <div class="container">
-    <!-- Available rooms -->
-    <h1 style="margin-top: 30px">Available rooms</h1>
+  <div class="box">
+    <h1>Available rooms</h1>
     <div class="list-group">
       <ul>
         <div v-for="(room, index) in userRooms" :key="index">
@@ -99,6 +98,19 @@ main {
 input[type="submit"]:hover {
   background-color: white;
   color: #703ab8;
+}
+
+.box {
+  background-color: white;
+  border: none;
+  font-weight: bold;
+  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  border-radius: 3px;
+  padding: 15px;
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
+  margin-top: 10px;
+  margin: 10px;
+  border: 2px solid #703ab8;
 }
 
 .error-paragraf {
