+++ frontend/src/components/ChatNew/ChatCreateRoom.vue
@@ -105,7 +105,7 @@ export default {
 </script>
 
 <template>
-  <div class="container">
+  <div class="box">
     <h1>Room creation</h1>
     <div>
       <input type="text" v-model="newRoomName" placeholder="Room Name" />
@@ -135,7 +135,7 @@ export default {
         This room will be private (only you and added users can see it).
       </div>
       <button class="submit-new-room new-room-button" @click="createRooms">
-        Créer {{ newRoomName }}
+        Create {{ newRoomName }}
       </button>
     </div>
   </div>
@@ -153,6 +153,19 @@ input {
   border-radius: 13px;
   font-weight: bold;
   font-size: 16px;
+}
+
+.box {
+  background-color: white;
+  border: none;
+  font-weight: bold;
+  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  border-radius: 3px;
+  padding: 15px;
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
+  margin-top: 10px;
+  margin: 10px;
+  border: 2px solid #703ab8;
 }
 
 .new-room-button {
